Ensure new Android feature READMEs are executable
This commit adds the new android-compile-dependencies feature hierarchy
to the multi-project build, and makes changes to their build scripts and
README files necessary to ensure that they (a) compile and (b) do not
produce diffs when running the `gradle readme` task.

Because these features are in an early draft stage, this means
essentially disabling most 'executable spec' functionality within them.
This is done by changing `   $ ` prompts to `    # ` so as to elide
being executed as a command when parsed by the `readme` task.

"Executable spec" headings have been renamed to "Usage".

Clean-up steps have been added at the end of Usage sections wherever
missing. Clean up steps are important in order to ensure that a usage
section can be executed any number of times while producing the same
result.