Move custom types to bottom of build scripts
It is often necessary to add custom classes and interfaces to a feature
build script in order to make it compile, or to generally express what
the build script is supposed to be able to do before we have properly
implemented what is necessary to *actually* do it.

By placing these types at the bottom of the build script, we help the
reader (whether users or fellow Gradle devs) focus on the build script
itself, which is the most important part. The custom types are in most
cases an implementation detail of the build script, and a temporary one
at that.