Simplify Travis CI configuration
Travis was originally set up to run on OS X (which requires the use of
language 'objective-c', and in anticipation of using tools like `hub` to
issue pull requests against the gradle/features repository.

These changes may come in the future, but they're not being used now and
therefore this configuration is being pruned.