Move the plugin code back to the top of the build script
This feature is about plugin development. Users will be focused on how
to develop a new software model plugin, and so the plugin code is the
most important part.

At some stage we'll probably pull the code into separate source files,
which will form the main part of the sample.