Enable standalone-junit-test-suite readme processing
This change updates commands in the usage section from using the no-op #
prompt to using the "live" $ prompt.

This means that when this project is processed using `gradle
useFeature`, these commands actually get run, and their live output is
written to the README file. If everything is accurate, no diff occurs.
If the output has for some reason, it causes a diff, indicating that the
feature is working differently than originally spec'ed.

Story: gradle/langos#103