Demonstrates a bug(?) in internal views.
I wanted to see if I could use an internal view to decorate existing binaries, in case it is a JUnit binary. So I introduced a `JUnitBinarySpec` that extends `JarBinarySpecInternal`, then declared this as an internal view for `JarBinarySpec`.
Then I replaced `JarBinarySpecInternal` in `createTestSuiteTask` with `JUnitBinarySpec`. As a consequence, `createTestSuiteTask` is no longer called. This looks like a bug.