Do not use fenced code blocks for diffs
Fenced code blocks are currently not processed by `gradle useFeature`.
Only four-space indented blocks are. This means that when directories
or files are renamed, the code-fenced diff output does not get updated
in the same way other commands do.

Fenced code blocks were used for diffs to ensure proper syntax
highlighting of diff output in the HTML rendered from the Markdown. This
is because fenced code blocks can take a 'language' parameter after the
first fence.

The solution to this problem is to update the `gradle useFeature` task
to be a bit smarter about the commands it encounters and to support
fenced code blocks as well as indented code blocks. But that can come
later.