Avoid CI build breakage due to "assemble runs tests" bug
This commit comments out two specs to avoid breaking the CI build.
Travis always runs `./gradlew assemble`, and as documented in
the bug referenced below, `assemble` currently triggers execution of
tests. Thes specs can be re-included in the build when (a) that bug is
fixed, or (b) support for the specs is actually implemented, whichever
comes first.

Bug: gradle/langos#118