Update jvm/test-execution feature spec
 - Use GFM task lists
 - Use GFM issue references to feature card and related stories

This feature spec is now the definitive source of information about
which stories need to be completed for this feature and in what order.
That is, there is no longer a task list in the feature *card*, only in
the feature *spec*. Some stories listed here in the feature spec do not
yet have a corresponding card on the board. These can be created as
necessary, i.e. when we get to them.

Feature: gradle/langos#102