Override Travis CI default 'install' command
This commit introduces a `gradle noop` task that is called during Travis
CI's 'install' phase [1]. As its name suggests, the task itself does
nothing, but serves to trigger downloading of the wrapper ahead of
calling `gradle useFeature`. This ensures that the Gradle wrapper
download console output does not show up in executable readme output.

Previously, `gradle assemble` was being called during the Travis CI's
'install' phase, as it is the default [2]. This had the effect of not
only eagerly downloading the wrapper (a good thing), but also eagerly
downloading any dependencies (usually a good thing for a normal Gradle
build, but undesirable in this case, since we want the executable spec
output to include dependency download console output.

[1]: https://docs.travis-ci.com/user/customizing-the-build/#Customizing-the-Installation-Step
[2]: https://docs.travis-ci.com/user/languages/java/#Dependency-Management