Capture dependency download logging in usage output
This change ensures that each gradle command executed in an spec's usage
section has --gradle-user-home set to a temporary directory created
specifically for that spec. The reason for doing this is to ensure that
dependency download console output is captured correctly as per what the
user would see if they were working against a clean cache and/or had
never resolved those dependencies before.

This change was prompted by noticing that dependency download console
output was being captured in the Travis CI build [1], where every build
happens against a clean cache.

This problem hadn't been noticed previously because Travis runs `gradle
assemble` as part of its 'install' lifecycle. In the case of JVM
libraries, assemble causes dependencies to be resolved, and prevents
this output from being captured in executable usage output during
`gradle useFeature`. In the case of test suites in the recently
introduced `testSuites {}` container, assemble does not trigger
dependency resolution, meaning that it happens on the fly during
execution of the `gradle useFeature` task, and therefore ended up
creating the diffs seen at [1] below.

As of a previous commit, `gradle assemble` is no longer run during the
Travis CI 'install' phase. Now, regardless the component or test suite
being built, dependencies are downloaded every time for every project,
and are guaranteed to show up in executable usage section output as the
user would see it if they were working against a fresh cache.

These changes have the negative effect of slowing the build
considerably, but this is done in exchange for the positive effect of
ensuring complete and deterministic build output.

[1]: https://travis-ci.org/gradle/features/builds/99266018