Do not run `gradle clean` in specs unless useful
Prior to this commit, most feature specs ran `gradle clean` as part of
the first step of the usage section to ensure that the project is indeed
clean and that usage output would therefore be predictable. This works,
but has the effect of creating extra steps for the spec author and noise
for the reader.

This commit removes all such unecessary invocations of `gradle clean` in
favor of simply calling `gradle clean useFeature` at the command line.
Note that the Travis CI build always runs against a clean environment,
so calling clean there is unnecessary. Nevertheless, the recently added
`noop` task has been removed in this commit in favor of invoking `gradle
clean` during the Travis CI 'install' phase. This is done to help
communicate that it's best to call `gradle clean useFeature` in any
case.